Move from Postgres to SQLite.
Mostly for ease of deployment. This is never meant to be used by
hundreds of people concurrently anyway.

+++ src/sql/create-raffle-entries.sql
@@ -1,7 +1,7 @@
 create table raffle_entries
 (
-    raffle_id bigint not null references raffles(id),
-    participant bigint not null,
+    raffle_id integer not null references raffles(id),
+    participant integer not null,
     chosen boolean not null default false,
 
     primary key (raffle_id, participant)
